+++ src/slices/playerSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../store';
+import { RootState } from '../store';
 
 export interface PlayerState {
   volume: number;
